Moved modal into its own class. Completed OrderDetails component layout.
