+++ src/components/modal/modal.jsx
@@ -1,0 +1,35 @@
+import { useState, useRef, useEffect } from "react";
+import styles from "./modal.module.css";
+import { CloseIcon } from "@ya.praktikum/react-developer-burger-ui-components";
+import ModalOverlay from "../modaloverlay/modaloverlay";
+
+const Modal = ({ children, onClose }) => {
+  const [closeIconState, setCloseIconState] = useState("primary");
+
+  const handleCloseButtonHover = () => {
+    setCloseIconState("secondary");
+  };
+
+  const handleCloseButtonLeave = () => {
+    setCloseIconState("primary");
+  };
+
+  return (
+    <ModalOverlay>
+      <div className={styles.container}>
+        <button
+          type="button"
+          className={styles.closeButton}
+          onClick={onClose}
+          onMouseOver={handleCloseButtonHover}
+          onMouseLeave={handleCloseButtonLeave}
+        >
+          <CloseIcon type={closeIconState} />
+        </button>
+        {children}
+      </div>
+    </ModalOverlay>
+  );
+};
+
+export default Modal;
