+++ src/components/modal/modal.module.css
@@ -1,0 +1,24 @@
+.container {
+  position: relative;
+  background: #1c1c21;
+  width: 720px;
+  min-height: 200px;
+  border: 1px solid rgba(76, 76, 255, 0.2);
+  box-shadow: 0px 24px 32px rgba(0, 0, 0, 0.04),
+    0px 16px 24px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04),
+    0px 0px 1px rgba(0, 0, 0, 0.04);
+  border-radius: 40px;
+}
+
+.closeButton {
+  margin: 0;
+  padding: 0;
+  position: absolute;
+  top: 60px;
+  right: 40px;
+  width: 24px;
+  height: 24px;
+  background: none;
+  border: none;
+  cursor: pointer;
+}
