+++ src/components/modaloverlay/modaloverlay.jsx
@@ -1,36 +1,11 @@
-import { createPortal } from "react-dom";
 import { useState, useRef, useEffect } from "react";
+import { createPortal } from "react-dom";
 import styles from "./modaloverlay.module.css";
-import { CloseIcon } from "@ya.praktikum/react-developer-burger-ui-components";
-
-const modalRoot = document.getElementById("react-modals");
 
 const ModalOverlay = ({ children, onClose }) => {
-  const [closeIconState, setCloseIconState] = useState("primary");
-
-  const handleCloseButtonHover = () => {
-    setCloseIconState("secondary");
-  };
-
-  const handleCloseButtonLeave = () => {
-    setCloseIconState("primary");
-  };
-
+  const modalRoot = document.getElementById("react-modals");
   return createPortal(
-    <div className={styles.overlay}>
-      <div className={styles.container}>
-        <button
-          type="button"
-          className={styles.closeButton}
-          onClick={onClose}
-          onMouseOver={handleCloseButtonHover}
-          onMouseLeave={handleCloseButtonLeave}
-        >
-          <CloseIcon type={closeIconState} />
-        </button>
-        {children}
-      </div>
-    </div>,
+    <div className={styles.overlay}>{children}</div>,
     modalRoot
   );
 };
