+++ src/components/orderdetails/orderdetails.jsx
@@ -1,13 +1,20 @@
 import styles from "./orderdetails.module.css";
+import orderAcceptedIcon from "../../images/order_accepted.svg";
 
 const OrderDetails = (props) => {
   return (
-    <div className={styles.container}>
-      <p className="text text_type_digits-large">{props.data.orderId}</p>
-      <p className="text text_type_main-medium">идентификатор заказа</p>
-      <img src="/order-accepted.svg" alt="Order accepted" />
+    <div className={styles.container + " mt-30 mb-30"}>
+      <p className={styles.orderid + " text text_type_digits-large"}>
+        {props.data.orderId}
+      </p>
+      <p className="text text_type_main-medium mt-8">идентификатор заказа</p>
+      <img
+        className="mt-15 mb-15"
+        src={orderAcceptedIcon}
+        alt="Order accepted"
+      />
       <p className="text text_type_main-small">Ваш заказ начали готовить</p>
-      <p className="text text_type_main-small text_color_inactive">
+      <p className="text text_type_main-small text_color_inactive mt-2">
         Дождитесь готовности на орбитальной станции
       </p>
     </div>
